// frontend/src/api/model/AdminAuditLogDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Instant } from "./Instant";

export interface AdminAuditLogDto {
  publicId: string;
  tableName: string;
  action: string;
  /** @nullable */
  changedColumns: string[] | null;
  summary: string;
  /** @nullable */
  performedByUserPublicId: string | null;
  performedBySystem: boolean;
  createdAt: Instant;
}